Loading data, and doing analysis on them
Just want to upload this for now, in case it is at all useful. I have essentially downloaded the 4 training files from "https://zenodo.org/records/1489920" (training, validation, training_ground, validation_ground) and this file organizes and tries to do analysis on them (uses a pre-trained model, and by training it fine tunes the model's parameters to fit this dataset). The comments in this file should be quite self-explanatory. But, I tried to run this on my computer, runs into memory issues and such as the files are quite large, but in case this is useful to anyone, im still going to upload this.